requirements.txt
python3-discogs-client===2.6